moved general gcp bucket interaction functions to a new dir
